+++ php/2fa.php
@@ -2,6 +2,7 @@
 namespace SIM\LOGIN;
 use SIM;
 use RobThree\Auth\TwoFactorAuth;
+use RobThree\Auth\Providers\Qr\BaconQrCodeProvider;
 use BaconQrCode\Renderer\ImageRenderer;
 use BaconQrCode\Renderer\Image\ImagickImageBackEnd;
 use BaconQrCode\Renderer\RendererStyle\RendererStyle;
@@ -27,7 +28,7 @@ if(!class_exists('RobThree\Auth\TwoFactorAuth')){
 function setupTimeCode(){
     $user                           = wp_get_current_user();
     $userId                         = $user->ID;
-    $twofa                          = new TwoFactorAuth();
+    $twofa                          = new TwoFactorAuth(new BaconQrCodeProvider());
     $setupDetails                   = new stdClass();
     $setupDetails->secretKey        = $twofa->createSecret();
 
@@ -157,7 +158,8 @@ function authenticate( $user) {
         
         //we did a succesfull webauthn or are on localhost
         if(
-            $_SERVER['HTTP_HOST'] == 'localhost' || 
+            $_SERVER['HTTP_HOST'] == 'localhost'  || 
+            str_contains($_SERVER['HTTP_HOST'], '.local') || 
             in_array('webauthn', $methods) && $_SESSION['webauthn'] == 'success'){
             //succesfull webauthentication done before
         }elseif(in_array('authenticator', $methods)){